pyproject.toml
[build-system]
requires = ["hatchling>=1.0.0"]
build-backend = "hatchling.build"

[project]
name = "ingest-anything"
version = "1.1.0"
authors = [
    { name="Clelia (Astra) Bertelli", email="astraberte9@gmail.com" },
]
description = "ingest-anything: from data to vector database effortlessly"
readme = "README.md"
requires-python = ">=3.10,<3.13"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    'chonkie[all]',
    'pydantic',
    'pdfitdown',
    'llama-index-readers-docling>0.1.0',
]

[project.urls]
Homepage = "https://github.com/AstraBert/ingest-anything"
Issues = "https://github.com/AstraBert/ingest-anything/issues"

[tool.setuptools.packages.find]
where = ["src"]
include = ["ingest_anything*"]

[tool.hatch.build.targets.wheel]
only-include = ["src/ingest_anything"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.hatch.envs.default]
dependencies = [
    "pytest",
    "llama-index-vector-stores-milvus>0.1.2",
    "llama-index-vector-stores-qdrant>0.1.3",
    "llama-index-vector-stores-weaviate>0.1",
    "llama-index-llms-mistralai>0.1",
]

[tool.hatch.envs.default.scripts]
test = '''
find src/ingest_anything -name "*.py" ! -name "__init__.py" -exec cp {} tests/ \; && \
pytest tests/*.py -p no:warnings && \
find tests/ -name "*.py" ! -name "*test*.py" -exec rm {} \;
'''

[tool.hatch.envs.default.py-version]
40 = "3.10"
41 = "3.11"
42 = "3.12"